+++ mobile/app/src/main/java/com/example/test/CartProduct.kt
@@ -1,7 +1,7 @@
 package com.example.test
 
 class CartProduct(var product: Product, var quantity: Int) {
-    fun getTotalPrice(): Int {
+    fun getTotalPrice(): Float {
         return product.price * quantity
     }
 }
