+++ mobile/app/src/main/java/com/example/test/Product.kt
@@ -1,3 +1,3 @@
 package com.example.test
 
-class Product(var name: String?, var price: Int, var description: String?, var image: String)
+class Product(var name: String?, var price: Float, var description: String?, var image: String)
